Adding support for DELETE after POST GET resources for full-circle tests.

+++ main.py
@@ -5,7 +5,11 @@ import json
 from tests import TESTS
 from config import BASE_URL
 
-def print_test(url, method, result, test_name, wanted, status_code, data, text, minimal=False):
+def print_test(url, method, result, test_name, wanted, status_code, data, text, minimal=False, mode=None, passed=None,only_200=None):
+    if only_200 == True:
+        if passed == True:
+            if status_code != 200:
+                return
     try:
         text = json.dumps(json.loads(text), indent=4)
     except:
@@ -66,7 +70,7 @@ def print_summary(tests_run, tests_passed, tests_failed):
     \033[91mFAILED:         %s \033[0m
         """ % (tests_run, tests_passed, tests_failed))
 
-def run_method_status(chosen_label, chosen_method, chosen_status, minimal=True):
+def run_method_status(chosen_label, chosen_method, chosen_status, minimal=True, mode=None,only_200=None):
     
     # Count tests run.
     tests_run = 0
@@ -125,11 +129,28 @@ def run_method_status(chosen_label, chosen_method, chosen_status, minimal=True):
                     text = None
 
                     if method == 'POST':
+
+                        # Check mode for 'all' run.
+                        if mode == 'ONLY_DELETE':
+                            continue
+
                         response = requests.post(url, data, cookies=cookies)
+
                     elif method == 'GET':
+
+                        # Check mode for 'all' run.
+                        if mode == 'ONLY_DELETE':
+                            continue
+
                         response = requests.get(url, data, cookies=cookies)
+
                     elif method == 'DELETE':
+
+                        if mode == 'SKIP_DELETE':
+                            continue
+
                         response = requests.delete(url, cookies=cookies)
+
                     else:
                         raise Exception("Unknown HTTP request method '%s'." % method)
 
@@ -138,12 +159,12 @@ def run_method_status(chosen_label, chosen_method, chosen_status, minimal=True):
 
                     if str(status_code) == status:
                         print_test(url, method, '\033[92mPASSED\033[0m', test_name, status, status_code, data, text,
-                                minimal=minimal)
+                                minimal=minimal, mode=mode, passed=True,only_200=only_200)
                         tests_run += 1
                         tests_passed += 1
                     else:
                         print_test(url, method, '\033[91mFAILED\033[0m', test_name, status, status_code, data, text,
-                                minimal=minimal)
+                                minimal=minimal, mode=mode, passed=True,only_200=only_200)
                         tests_run += 1
                         tests_failed += 1
 
@@ -156,10 +177,11 @@ def run_method_status(chosen_label, chosen_method, chosen_status, minimal=True):
 
 # Run all of the tests with this label (for all specified
 # HTTP methods and HTTP status codes). 
-def run_label(label, minimal=True):
+def run_label(label, minimal=True, mode=None,only_200=None):
     tests_run = 0
     tests_failed = 0
     tests_passed = 0
+    failed_labels = {}
     tests = list(filter(lambda test:
         test.get('label') == label, TESTS))
     for test in tests:
@@ -167,13 +189,17 @@ def run_label(label, minimal=True):
             if method in ['POST','GET','DELETE']:
                 for status, test in val.items():
                     _tests_run, _tests_passed, _tests_failed = run_method_status(
-                        label, method, status, minimal=True)
+                        label, method, status, minimal=True, mode=mode,only_200=only_200)
                     tests_run += _tests_run
                     tests_passed += _tests_passed
                     tests_failed += _tests_failed
+                    if _tests_failed > 0:
+                        if label not in failed_labels:
+                            failed_labels[label] = []
+                        failed_labels[label].append({method: status})
     if minimal == False:
         print_summary(tests_run, tests_passed, tests_failed)
-    return tests_run, tests_passed, tests_failed
+    return tests_run, tests_passed, tests_failed, failed_labels
 
 # All the tests I've already done.
 DONE_LIST = [
@@ -322,6 +348,22 @@ Examples:
       $ python3 main.py [label]""")
     exit()
 
+def dump_spec():
+    # Dump the 200s for the spec.
+    spec = {}
+    for test in TESTS:
+        spec[test['url']] = {}
+        if test.get('POST') is not None:
+            spec[test['url']]['POST'] = test.get('POST').get('200')
+        if test.get('GET') is not None:
+            spec[test['url']]['GET'] = test.get('GET').get('200')
+        if test.get('DELETE') is not None:
+            spec[test['url']]['DELETE'] = test.get('DELETE').get('200')
+    string = json.dumps(spec, indent=4)
+    with open('spec.json','w') as f:
+        f.write(string)
+        f.close()
+
 if __name__ == '__main__':
 
     print(chr(27) + "[2J")
@@ -334,17 +376,7 @@ if __name__ == '__main__':
     label = sys.argv[1]
 
     if label == 'dumpspec':
-        # Dump the 200s for the spec.
-        spec = {}
-        for test in TESTS:
-            spec[test['url']] = {}
-            if test.get('POST') is not None:
-                spec[test['url']]['POST'] = test.get('POST').get('200')
-            if test.get('GET') is not None:
-                spec[test['url']]['GET'] = test.get('GET').get('200')
-            if test.get('DELETE') is not None:
-                spec[test['url']]['DELETE'] = test.get('DELETE').get('200')
-        print(json.dumps(spec, indent=4))
+        dump_spec()
         exit()
     if label == 'dumptests':
         # Dump the all the tests.
@@ -356,12 +388,41 @@ if __name__ == '__main__':
         tests_run = 0
         tests_failed = 0
         tests_passed = 0
+        failed_labels = []
+        only_200 = False
+       
+        DONE_LIST = list(filter(lambda lab: lab[0:7] == 'cir_pag', DONE_LIST))
+        """
+        only_200 = True 
+        """
+
+        # Run the labels skipping DELETE
         for _label in DONE_LIST:
-            _tests_run, _tests_passed, _tests_failed = run_label(_label)
+            _tests_run, _tests_passed, _tests_failed, _failed_labels = run_label(_label, mode='SKIP_DELETE', only_200=only_200)
             tests_run += _tests_run
             tests_passed += _tests_passed
             tests_failed += _tests_failed
+            for _failed_label, methods in _failed_labels.items():
+                failed_labels.append({_failed_label: methods})
+        # Run the labels for only DELETE
+        for _label in DONE_LIST:
+            _tests_run, _tests_passed, _tests_failed, _failed_labels = run_label(_label, mode='ONLY_DELETE', only_200=only_200)
+            tests_run += _tests_run
+            tests_passed += _tests_passed
+            tests_failed += _tests_failed
+            for _failed_label, methods in _failed_labels.items():
+                failed_labels.append({_failed_label: methods})
+        # Summary
         print_summary(tests_run, tests_passed, tests_failed)
+        if len(failed_labels) > 0:
+            print("Failed Tests:")
+            with open('failed_tests.txt','w') as f:
+                string = json.dumps(failed_labels, indent=4)
+                f.write(string)
+                f.close()
+                print(string)
+        # Dump the spec with the live working variables.
+        dump_spec()
         exit()
     # Not specifying an HTTP method or status will run all tests with that label.
     if len(sys.argv) == 2:
